+++ actions/gha-workflow-validator/src/__tests__/utils.test.ts
@@ -1,28 +1,25 @@
-import { filterForRelevantChanges, parseGithubDiff } from "../utils.js";
+import {
+  filterForRelevantChanges,
+  parseGithubDiff,
+  doValidationErrorsExist,
+  processLineValidationResults,
+} from "../utils.js";
 import { GithubFiles, getComparison } from "../github.js";
 import { getNock, getTestOctokit } from "./__helpers__/test-utils.js";
 
 import { vi, describe, it, expect } from "vitest";
+import {
+  FileValidationResult,
+  LineValidationResult,
+  ValidationMessage,
+  ValidationType,
+} from "../validations/validation-check.js";
 
 const nockBack = getNock();
 
-vi.mock("@actions/core", () => ({
-  setFailed: (msg: string) => {
-    console.log(`setFailed (stub): ${msg}`);
-  },
-  error: (msg: string) => {
-    console.log(`error (stub): ${msg}`);
-  },
-  warning: (msg: string) => {
-    console.log(`warn (stub): ${msg}`);
-  },
-  info: (msg: string) => {
-    console.log(`info (stub): ${msg}`);
-  },
-  debug: (msg: string) => {
-    console.log(`debug (stub): ${msg}`);
-  },
-}));
+vi.mock("@actions/core", async () => {
+  return (await import("./__helpers__/test-utils.js")).coreLoggingStubs();
+});
 
 const simplePatchResponse: GithubFiles = [
   {
@@ -161,5 +158,433 @@ describe(parseGithubDiff.name, () => {
     expect(parsedFiles).toMatchSnapshot();
 
     nockDone();
+  });
+});
+
+describe(doValidationErrorsExist.name, () => {
+  it("should return true with no line validations", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(false);
+  });
+
+  it("should return true with only warning validations", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "warning",
+              },
+            ],
+          },
+        ],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(false);
+  });
+
+  it("should return true with only ignored validations", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "warning",
+              },
+            ],
+          },
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "ignored",
+              },
+            ],
+          },
+        ],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(false);
+  });
+
+  it("should return true with warning and ignored validations", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "ignored",
+              },
+            ],
+          },
+        ],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(false);
+  });
+
+  it("should return false with single error", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "error",
+              },
+            ],
+          },
+        ],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(true);
+  });
+
+  it("should return false with all types", () => {
+    const fileValidations: FileValidationResult[] = [
+      {
+        filename: ".github/workflows/test.yml",
+        lineValidations: [
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "error",
+              },
+            ],
+          },
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "warning",
+              },
+            ],
+          },
+          {
+            filename: ".github/workflows/test.yml",
+            line: {
+              lineNumber: 1,
+              content: "line 1",
+              operation: "add",
+              ignored: false,
+            },
+            messages: [
+              {
+                message: "Error",
+                type: ValidationType.VERSION_COMMENT,
+                severity: "ignored",
+              },
+            ],
+          },
+        ],
+      },
+    ];
+    const result = doValidationErrorsExist(fileValidations);
+    expect(result).toBe(true);
+  });
+});
+
+describe(processLineValidationResults.name, () => {
+  it("should combine empty results", () => {
+    const results = processLineValidationResults([]);
+    expect(results).toEqual([]);
+  });
+
+  it("should combine single result", () => {
+    const singleLineValidation: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.VERSION_COMMENT,
+          severity: "error",
+        },
+      ],
+    };
+    const results = processLineValidationResults([singleLineValidation]);
+    expect(results).toEqual([singleLineValidation]);
+  });
+
+  it("should combine two results for the same line", () => {
+    const lineOneValidationOne: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.VERSION_COMMENT,
+          severity: "error",
+        },
+      ],
+    };
+
+    const lineOneValidationTwo: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.NODE_VERSION,
+          severity: "warning",
+        },
+      ],
+    };
+
+    const results = processLineValidationResults([
+      lineOneValidationOne,
+      lineOneValidationTwo,
+    ]);
+    expect(results.length).toBe(1);
+
+    const combinedLineValidation = results[0];
+    expect(combinedLineValidation.messages).toEqual([
+      ...lineOneValidationOne.messages,
+      ...lineOneValidationTwo.messages,
+    ]);
+  });
+
+  it("should combine two results for two lines (and sort by line)", () => {
+    const lineTwoValidationOne: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 2,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.VERSION_COMMENT,
+          severity: "error",
+        },
+      ],
+    };
+
+    const lineTwoValidationTwo: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 2,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.NODE_VERSION,
+          severity: "warning",
+        },
+      ],
+    };
+
+    const lineOneValidationOne: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.VERSION_COMMENT,
+          severity: "error",
+        },
+      ],
+    };
+
+    const lineOneValidationTwo: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: false,
+      },
+      messages: [
+        {
+          message: "Error",
+          type: ValidationType.NODE_VERSION,
+          severity: "warning",
+        },
+      ],
+    };
+
+    const results = processLineValidationResults([
+      lineOneValidationOne,
+      lineOneValidationTwo,
+      lineTwoValidationOne,
+      lineTwoValidationTwo,
+    ]);
+    expect(results.length).toBe(2);
+
+    const line1LV = results[0];
+    expect(line1LV.messages).toEqual([
+      ...lineOneValidationOne.messages,
+      ...lineOneValidationTwo.messages,
+    ]);
+
+    const line2LV = results[0];
+    expect(line2LV.messages).toEqual([
+      ...lineTwoValidationOne.messages,
+      ...lineTwoValidationTwo.messages,
+    ]);
+  });
+
+  it("should lower severity for ignored lines", () => {
+    const message: ValidationMessage = {
+      message: "Error",
+      type: ValidationType.VERSION_COMMENT,
+      severity: "error",
+    };
+
+    const singleLineValidation: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: true,
+      },
+      messages: [message],
+    };
+    const results = processLineValidationResults([singleLineValidation]);
+    expect(results).toEqual([
+      {
+        ...singleLineValidation,
+        messages: [{ ...message, severity: "ignored" }],
+      },
+    ]);
+  });
+
+  it("should lower severity for ignored lines but not new ignore comments", () => {
+    const ignoresCommentMessage: ValidationMessage = {
+      message: "Error",
+      type: ValidationType.IGNORE_COMMENT,
+      severity: "error",
+    };
+
+    const lowerableMessage: ValidationMessage = {
+      message: "Error",
+      type: ValidationType.VERSION_COMMENT,
+      severity: "error",
+    };
+
+    const singleLineValidation: LineValidationResult = {
+      filename: ".github/workflows/test.yml",
+      line: {
+        lineNumber: 1,
+        content: "line 1",
+        operation: "add",
+        ignored: true,
+      },
+      messages: [ignoresCommentMessage, lowerableMessage],
+    };
+    const results = processLineValidationResults([singleLineValidation]);
+    expect(results).toEqual([
+      {
+        ...singleLineValidation,
+        messages: [
+          ignoresCommentMessage,
+          { ...lowerableMessage, severity: "ignored" },
+        ],
+      },
+    ]);
   });
 });